ui: Move pagination integration tests into separate module
